feat(solver): vendored CombatSolver 0.27→0.28 engine + RitsuLib IL + GameRef private access
Adopt upstream 0.28 engine verbatim (Engine/Prediction/Search + needed Runtime data
types) for trustworthy current-turn hand-grounding (cross-turn/cycle/budget planning),
and restored RitsuLib card-effect fidelity (HarmonyIl, IComputedDynamicVar, GetMaxHandSize)
which the port's self-built IlReader had broken (unsupported BODY_SLAM/DRUM_OF_BATTLE/POMMEL_STRIKE).

- csproj: Krafs.Publicizer (compile-time publicize) + STS2-RitsuLib reference.
- Re-apply GameRef reflection to all direct `MegaCrit.Sts2.Core.*` private accesses
  (Krafs is compile-time only; RitsuLib runtime-publicization isn't available to a
  mods/-dir mod, so reflection reads the game's private members).
- Keep the port's lean facade (CombatSolverRuntime/EntryShim/CombatSolverFacade), the
  co-op single-player-guard removal, and declare RitsuLib as a mod dependency.
- IlReader.cs (self-built IL decoder) dropped in favor of upstream HarmonyIl.

+++ game_mod/neko_comm/Solver/Engine/InCombat/Mirrors/Hooks/Card/AfterCardDrawnMirrors.cs
@@ -218,7 +218,7 @@ internal static class AfterCardDrawnMirrors
             var previewCard = (KinglyPunch)context.MutablePreviewCard;
             var damageIncrease = previewCard.DynamicVars[GameRef.GetStatic<string>(typeof(KinglyPunch), "_increaseKey")].BaseValue;
             previewCard.DynamicVars.Damage.BaseValue += damageIncrease;
-            GameRef.Set(previewCard, "ExtraDamage", GameRef.Get<int>(previewCard, "ExtraDamage") + damageIncrease);
+            GameRef.Set(previewCard, "ExtraDamage", GameRef.Get<decimal>(previewCard, "ExtraDamage") + damageIncrease);
         }
     }
